People Flow Controller Totem
Engineering project developed in 2020 for the Sesi Lightning Challenge - Back to School competition. This competition, my team won a place to participate in her national stage.
The project we elaborated constitutes a totem that, with the use of an arduino board, infrared sensors and an LCD display, were able to prevent people from gathering in closed places. If you're interested, you can read the project and find out more about it.